move the doctrine:fixtures:load command out of the Command folder to allow for loading any future commands which might be added
